+++ src/render/shader/xr-yg-zb.vs
@@ -8,5 +8,5 @@ out vec3 PosColor;
 void main()
 {
     gl_Position = vec4(Position, 1.0);
-    PosColor = (Position + 1) / 2;// move range from [-1, 1] to [0, 0.5]
+    PosColor = (Position + 1) / 2;// move range from [-1, 1] to [0, 1]
 }
